Add password reset templates (four of them) ...
... This commit now has moved away from using templates based
on django admin. The templates for changing a password as well
as resetting a password are all now based on the project's base.html.

All templates allow for i18n.

These views do not really require testing as they are testing as part of
the django contib.auth app.

The example of login.html was used to hardcode the forms (they are
relatively simple) in order to style them with bootstrap classes.

We could consider moving towards a templatetag as `add_css_class`
to add a css class to an input or a label.