Upping max number of exportable pages to 10000.
This is to remedy a regression brought about by the recent refactoring
that changed submitting the list of selected pages by JSON to submitting
it as a POST request. The use case of wanting to export all pages at
once began failing silently for documents >1000 pages (Django's
default).

This change has so far only been tested on Bibliotheca Arabica's
instance where the reverse proxy is haproxy and not nginx!